+++ hostfunctions/dgraph.go
@@ -8,26 +8,25 @@ import (
 	"context"
 	"fmt"
 
-	"hmruntime/dqlclient"
+	"hmruntime/dgraphclient"
 	"hmruntime/logger"
 
 	wasm "github.com/tetratelabs/wazero/api"
 )
 
 func init() {
-	addHostFunction("executeDQLQuery", hostExecuteDQLQuery, withI32Params(3), withI32Result())
-	addHostFunction("executeDQLMutations", hostExecuteDQLMutations, withI32Params(3), withI32Result())
+	addHostFunction("executeDQL", hostExecuteDQL, withI32Params(2), withI32Result())
 	addHostFunction("dgraphAlterSchema", hostDgraphAlterSchema, withI32Params(2), withI32Result())
 	addHostFunction("dgraphDropAttr", hostDgraphDropAttr, withI32Params(2), withI32Result())
 	addHostFunction("dgraphDropAll", hostDgraphDropAll, withI32Params(1), withI32Result())
 }
 
-func hostExecuteDQLQuery(ctx context.Context, mod wasm.Module, stack []uint64) {
+func hostExecuteDQL(ctx context.Context, mod wasm.Module, stack []uint64) {
 
 	// Read input parameters
-	var hostName, query, varsJson string
-	var mutations []string
-	if err := readParams(ctx, mod, stack, &hostName, &query, &varsJson); err != nil {
+	var hostName string
+	var req dgraphclient.Request
+	if err := readParams(ctx, mod, stack, &hostName, &req); err != nil {
 		logger.Err(ctx, err).Msg("Error reading input parameters.")
 		return
 	}
@@ -38,13 +37,13 @@ func hostExecuteDQLQuery(ctx context.Context, mod wasm.Module, stack []uint64) {
 		Completed: "Completed DQL operation.",
 		Cancelled: "Cancelled DQL operation.",
 		Error:     "Error executing DQL operation.",
-		Detail:    fmt.Sprintf("Host: %s Query: %s Mutations: %v", hostName, query, mutations),
+		Detail:    fmt.Sprintf("Host: %s Req: %s", hostName, fmt.Sprint(req)),
 	}
 
 	// Prepare the host function
-	var result string
+	var result *dgraphclient.Response
 	fn := func() (err error) {
-		result, err = dqlclient.ExecuteQuery(ctx, hostName, query, varsJson)
+		result, err = dgraphclient.Execute(ctx, hostName, &req)
 		return err
 	}
 
@@ -57,43 +56,7 @@ func hostExecuteDQLQuery(ctx context.Context, mod wasm.Module, stack []uint64) {
 	if err := writeResults(ctx, mod, stack, result); err != nil {
 		logger.Err(ctx, err).Msg("Error writing results to wasm memory.")
 	}
-}
 
-func hostExecuteDQLMutations(ctx context.Context, mod wasm.Module, stack []uint64) {
-
-	// Read input parameters
-	var hostName string
-	var setMutations, delMutations []string
-	if err := readParams(ctx, mod, stack, &hostName, &setMutations, &delMutations); err != nil {
-		logger.Err(ctx, err).Msg("Error reading input parameters.")
-		return
-	}
-
-	// Prepare log messages
-	msgs := &hostFunctionMessages{
-		Starting:  "Executing DQL mutations.",
-		Completed: "Completed DQL mutations.",
-		Cancelled: "Cancelled DQL mutations.",
-		Error:     "Error executing DQL mutations.",
-		Detail:    fmt.Sprintf("Host: %s SetMutations: %v DelMutations: %v", hostName, setMutations, delMutations),
-	}
-
-	// Prepare the host function
-	var result map[string]string
-	fn := func() (err error) {
-		result, err = dqlclient.ExecuteMutations(ctx, hostName, setMutations, delMutations)
-		return err
-	}
-
-	// Call the host function
-	if ok := callHostFunction(ctx, fn, msgs); !ok {
-		return
-	}
-
-	// Write the results
-	if err := writeResults(ctx, mod, stack, result); err != nil {
-		logger.Err(ctx, err).Msg("Error writing results to wasm memory.")
-	}
 }
 
 func hostDgraphAlterSchema(ctx context.Context, mod wasm.Module, stack []uint64) {
@@ -117,7 +80,7 @@ func hostDgraphAlterSchema(ctx context.Context, mod wasm.Module, stack []uint64)
 	// Prepare the host function
 	var result string
 	fn := func() (err error) {
-		result, err = dqlclient.AlterSchema(ctx, hostName, schema)
+		result, err = dgraphclient.AlterSchema(ctx, hostName, schema)
 		return err
 	}
 
@@ -153,7 +116,7 @@ func hostDgraphDropAttr(ctx context.Context, mod wasm.Module, stack []uint64) {
 	// Prepare the host function
 	var result string
 	fn := func() (err error) {
-		result, err = dqlclient.DropAttr(ctx, hostName, attr)
+		result, err = dgraphclient.DropAttr(ctx, hostName, attr)
 		return err
 	}
 
@@ -189,7 +152,7 @@ func hostDgraphDropAll(ctx context.Context, mod wasm.Module, stack []uint64) {
 	// Prepare the host function
 	var result string
 	fn := func() (err error) {
-		result, err = dqlclient.DropAll(ctx, hostName)
+		result, err = dgraphclient.DropAll(ctx, hostName)
 		return err
 	}